requirements.txt
# Telegram Bot
aiogram>=3.3.0
aiohttp>=3.9.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Bitrix24 API
aiohttp>=3.9.0

# CSV Processing
pandas>=2.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Scheduling
apscheduler>=3.10.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
responses>=0.24.0

# Logging
python-json-logger>=2.0.0

# Utils
pydantic>=2.5.0
pydantic-settings>=2.1.0
